Test additions, improvements, refactoring.
