+++ azure-eventhubs-eph/src/test/java/com/microsoft/azure/eventprocessorhost/SmokeTest.java
@@ -10,6 +10,7 @@ import static org.junit.Assert.*;
 import java.time.Instant;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadPoolExecutor;
 
 import org.junit.Test;
@@ -21,12 +22,10 @@ import com.microsoft.azure.eventhubs.PartitionReceiver;
 
 public class SmokeTest extends TestBase
 {
-	/*
-	@Test
-	public void smokeTest() throws Exception
+	//@Test
+	public void SendRecv1MsgTest() throws Exception
 	{
-		PerTestSettings settings = new PerTestSettings("smoketest");
-		//settings.inBlobPrefix = "testprefix";
+		PerTestSettings settings = new PerTestSettings("SendRecv1Msg");
 		settings = testSetup(settings); 
 
 		settings.outUtils.sendToAny(settings.outTelltale);
@@ -39,7 +38,8 @@ public class SmokeTest extends TestBase
 	public void receiveFromNowTest() throws Exception
 	{
 		// Doing two iterations with the same "now" requires storing the "now" value instead of
-		// using the current time when the initial offset provider is executed.
+		// using the current time when the initial offset provider is executed. It also requires
+		// that the "now" be before the first send.
 		final Instant storedNow = Instant.now();
 
 		// Do the first iteration.
@@ -50,13 +50,13 @@ public class SmokeTest extends TestBase
 		// The purpose of running a second iteration is to look for bugs that occur when leases have been
 		// created and persisted but checkpoints have not, so it is vital that the second iteration uses the
 		// same storage container.
-		receiveFromNowIteration(storedNow, 2, 2, firstSettings.inoutStorageContainerName);
+		receiveFromNowIteration(storedNow, 2, 2, firstSettings.inoutEPHConstructorArgs.getStorageContainerName());
 	}
 	
 	private PerTestSettings receiveFromNowIteration(final Instant storedNow, int iteration, int expectedMessages, String containerName) throws Exception
 	{
-		PerTestSettings settings = new PerTestSettings("receiveFromNowTest-iter-" + iteration);
-		settings.inForcedContainerName = containerName;
+		PerTestSettings settings = new PerTestSettings("receiveFromNow-iter-" + iteration);
+		settings.inoutEPHConstructorArgs.setStorageContainerName(containerName);
 		settings.inOptions.setInitialOffsetProvider((partitionId) -> { return storedNow; });
 		settings = testSetup(settings);
 
@@ -67,18 +67,21 @@ public class SmokeTest extends TestBase
 		
 		return settings;
 	}
-	
-	//@Test
+
+	@Test
 	public void receiveFromCheckpoint() throws Exception
 	{
 		PerTestSettings firstSettings = receiveFromCheckpointIteration(1, SmokeTest.ANY_NONZERO_COUNT, null);
-		receiveFromCheckpointIteration(2, firstSettings.outPartitionIds.size(), firstSettings.inoutStorageContainerName);
+		System.out.println("Waiting five minutes");
+		Thread.sleep(1000*60*5);
+		System.out.println("Continuing");
+		receiveFromCheckpointIteration(2, firstSettings.outPartitionIds.size(), firstSettings.inoutEPHConstructorArgs.getStorageContainerName());
 	}
 
 	private PerTestSettings receiveFromCheckpointIteration(int iteration, int expectedMessages, String containerName) throws Exception
 	{
-		PerTestSettings settings = new PerTestSettings("receiveFromCkptTest-iter-" + iteration);
-		settings.inForcedContainerName = containerName;
+		PerTestSettings settings = new PerTestSettings("receiveFromCkpt-iter-" + iteration);
+		settings.inoutEPHConstructorArgs.setStorageContainerName(containerName);
 		settings.inDoCheckpoint = true;
 		settings = testSetup(settings);
 
@@ -96,8 +99,11 @@ public class SmokeTest extends TestBase
 	//@Test
 	public void receiveAllPartitionsTest() throws Exception
 	{
-		PerTestSettings settings = new PerTestSettings("receiveAllPartitionsTest");
-		settings.inOptions.setInitialOffsetProvider((partitionId) -> { return Instant.now(); });
+		// Save "now" to avoid race with sender startup.
+		final Instant savedNow = Instant.now();
+		
+		PerTestSettings settings = new PerTestSettings("receiveAllPartitions");
+		settings.inOptions.setInitialOffsetProvider((partitionId) -> { return savedNow; });
 		settings = testSetup(settings);
 
 		final int maxGeneration = 10;
@@ -105,14 +111,14 @@ public class SmokeTest extends TestBase
 		{
 			for (String id : settings.outPartitionIds)
 			{
-				settings.outUtils.sendToPartition(id, "receiveAllPartitionsTest-" + id + "-" + generation);
+				settings.outUtils.sendToPartition(id, "receiveAllPartitions-" + id + "-" + generation);
 			}
-			System.out.println("Generation " + generation + " sent");
+			TestUtilities.log("Generation " + generation + " sent\n");
 		}
 		for (String id : settings.outPartitionIds)
 		{
 			settings.outUtils.sendToPartition(id, settings.outTelltale);
-			System.out.println("Telltale " + id + " sent");
+			TestUtilities.log("Telltale " + id + " sent\n");
 		}
 		for (String id : settings.outPartitionIds)
 		{
@@ -123,201 +129,35 @@ public class SmokeTest extends TestBase
 	}
 	
 	//@Test
-	public void conflictingHosts() throws Exception
+	public void receiveAllPartitionsWithUserExecutorTest() throws Exception
 	{
-		System.out.println("conflictingHosts starting");
+		// Save "now" to avoid race with sender startup.
+		final Instant savedNow = Instant.now();
 		
-		RealEventHubUtilities utils = new RealEventHubUtilities();
-		utils.setup(-1);
-		
-		String telltale = "conflictingHosts-telltale-" + EventProcessorHost.safeCreateUUID();
-		String conflictingName = "conflictingHosts-NOTSAFE";
-		String storageName = conflictingName.toLowerCase() + EventProcessorHost.safeCreateUUID();
-		boolean doCheckpointing = false;
-		boolean doMarker = false;
-		
-		PrefabGeneralErrorHandler general1 = new PrefabGeneralErrorHandler();
-		PrefabProcessorFactory factory1 = new PrefabProcessorFactory(telltale, doCheckpointing, doMarker);
-		EventProcessorHost host1 = new EventProcessorHost(conflictingName, utils.getConnectionString().getEntityPath(),
-				utils.getConsumerGroup(), utils.getConnectionString().toString(),
-				TestUtilities.getStorageConnectionString(), storageName);
-		EventProcessorOptions options1 = EventProcessorOptions.getDefaultOptions();
-		options1.setExceptionNotification(general1);
-		
-		PrefabGeneralErrorHandler general2 = new PrefabGeneralErrorHandler();
-		PrefabProcessorFactory factory2 = new PrefabProcessorFactory(telltale, doCheckpointing, doMarker);
-		EventProcessorHost host2 = new EventProcessorHost(conflictingName, utils.getConnectionString().getEntityPath(),
-				utils.getConsumerGroup(), utils.getConnectionString().toString(),
-				TestUtilities.getStorageConnectionString(), storageName);
-		EventProcessorOptions options2 = EventProcessorOptions.getDefaultOptions();
-		options2.setExceptionNotification(general2);
+		PerTestSettings settings = new PerTestSettings("rcvAllPartsUserExec");
+		settings.inOptions.setInitialOffsetProvider((partitionId) -> { return savedNow; });
+		settings.inoutEPHConstructorArgs.setExecutor(Executors.newCachedThreadPool());
+		settings = testSetup(settings);
 
-		host1.registerEventProcessorFactory(factory1, options1);
-		host2.registerEventProcessorFactory(factory2, options2);
-		
-		int i = 0;
-		while (true)
+		final int maxGeneration = 10;
+		for (int generation = 0; generation < maxGeneration; generation++)
 		{
-			utils.sendToAny("conflict-" + i++, 10);
-			System.out.println("\n." + factory1.getEventsReceivedCount() + "." + factory2.getEventsReceivedCount() + ":" +
-					((ThreadPoolExecutor)EventProcessorHost.getExecutorService()).getPoolSize() + ":" +
-					Thread.activeCount());
-			//DebugThread.printThreadStatuses();
-			Thread.sleep(100);
+			for (String id : settings.outPartitionIds)
+			{
+				settings.outUtils.sendToPartition(id, "receiveAllPartitionsWithUserExecutor-" + id + "-" + generation);
+			}
+			TestUtilities.log("Generation " + generation + " sent\n");
 		}
-	}
-	*/
-	
-	//@Test
-	public void rawEpochStealing() throws Exception
-	{
-		RealEventHubUtilities utils = new RealEventHubUtilities();
-		utils.setup(-1);
-
-		int clientSerialNumber = 0;
-		while (true)
+		for (String id : settings.outPartitionIds)
 		{
-			Thread[] blah = new Thread[Thread.activeCount() + 10];
-			int actual = Thread.enumerate(blah); 
-			if (actual >= blah.length)
-			{
-				System.out.println("Lost some threads");
-			}
-			int parkedCount = 0;
-			String selectingList = "";
-			boolean display = true;
-			for (int i = 0; i < actual; i++)
-			{
-				display = true;
-				StackTraceElement[] bloo = blah[i].getStackTrace();
-				String show = "nostack";
-				if (bloo.length > 0)
-				{
-					show = bloo[0].getClassName() + "." + bloo[0].getMethodName();
-					if (show.compareTo("sun.misc.Unsafe.park") == 0)
-					{
-						parkedCount++;
-						display = false;
-					}
-					else if (show.compareTo("sun.nio.ch.WindowsSelectorImpl$SubSelector.poll0") == 0)
-					{
-						selectingList += (" " + blah[i].getId());
-						display = false;
-					}
-				}
-				if (display)
-				{
-					System.out.print(" " + blah[i].getId() + ":" + show);
-				}
-			}
-			System.out.println("\nParked: " + parkedCount + "  SELECTING: " + selectingList);
-			
-			System.out.println("Client " + clientSerialNumber + " starting");
-			EventHubClient client = EventHubClient.createFromConnectionStringSync(utils.getConnectionString().toString());
-			PartitionReceiver receiver =
-					client.createEpochReceiver(utils.getConsumerGroup(), "0", PartitionReceiver.START_OF_STREAM, 1).get();
-
-			boolean useReceiveHandler = false;
-			
-			if (useReceiveHandler)
-			{
-				Blah b = new Blah(clientSerialNumber++, receiver, client);
-				receiver.setReceiveHandler(b).get();
-				// wait for messages to start flowing
-				b.waitForReceivedMessages().get();
-			}
-			else
-			{
-				receiver.receiveSync(1);
-				System.out.println("Received a message");
-			}
-			
-			// Enable these lines to avoid overlap
-			/* */
-			try
-			{
-				System.out.println("Non-overlap close of PartitionReceiver");
-				if (useReceiveHandler)
-				{
-					receiver.setReceiveHandler(null).get();
-				}
-				receiver.close().get();
-			}
-			catch (InterruptedException | ExecutionException e)
-			{
-				System.out.println("Client " + clientSerialNumber + " failed while closing PartitionReceiver: " + e.toString());
-			}
-			try
-			{
-				System.out.println("Non-overlap close of EventHubClient");
-				client.close().get();
-			}
-			catch (InterruptedException | ExecutionException e)
-			{
-				System.out.println("Client " + clientSerialNumber + " failed while closing EventHubClient: " + e.toString());
-			}
-			System.out.println("Client " + clientSerialNumber + " closed");
-			/* */
-
-			System.out.println("Threads: " + Thread.activeCount());
+			settings.outUtils.sendToPartition(id, settings.outTelltale);
+			TestUtilities.log("Telltale " + id + " sent\n");
 		}
-	}
-	
-	private class Blah extends PartitionReceiveHandler
-	{
-		private int clientSerialNumber;
-		private PartitionReceiver receiver;
-		private EventHubClient client;
-		private CompletableFuture<Void> receivedMessages = null;
-		private boolean firstEvents = true;
-		
-		protected Blah(int clientSerialNumber, PartitionReceiver receiver, EventHubClient client)
+		for (String id : settings.outPartitionIds)
 		{
-			super(300);
-			this.clientSerialNumber = clientSerialNumber;
-			this.receiver = receiver;
-			this.client = client;
+			waitForTelltale(settings, id);
 		}
 		
-		CompletableFuture<Void> waitForReceivedMessages()
-		{
-			this.receivedMessages = new CompletableFuture<Void>();
-			return this.receivedMessages;
-		}
-
-		@Override
-		public void onReceive(Iterable<EventData> events)
-		{
-			if (this.firstEvents)
-			{
-				System.out.println("Client " + this.clientSerialNumber + " got events");
-				this.receivedMessages.complete(null);
-				this.firstEvents = false;
-			}
-		}
-
-		@Override
-		public void onError(Throwable error)
-		{
-			System.out.println("Client " + this.clientSerialNumber + " got " + error.toString());
-			try
-			{
-				this.receiver.close().get();
-			}
-			catch (InterruptedException | ExecutionException e)
-			{
-				System.out.println("Client " + this.clientSerialNumber + " failed while closing PartitionReceiver: " + e.toString());
-			}
-			try
-			{
-				this.client.close().get();
-			}
-			catch (InterruptedException | ExecutionException e)
-			{
-				System.out.println("Client " + this.clientSerialNumber + " failed while closing EventHubClient: " + e.toString());
-			}
-			System.out.println("Client " + this.clientSerialNumber + " closed");
-		}
-		
+		testFinish(settings, (settings.outPartitionIds.size() * (maxGeneration + 1))); // +1 for the telltales
 	}
 }
